Move the local-cluster exe code into the exe source dir
