--- lib/local-cluster/exe/Cluster.hs
@@ -1,6 +1,0 @@
-import qualified Cardano.Wallet.LocalCluster as LocalCluster
-
-import Prelude
-
-main :: IO ()
-main = LocalCluster.main
