+++ lib/local-cluster/local-cluster.cabal
@@ -51,6 +51,7 @@ library
     Cardano.Wallet.Launch.Cluster.Cluster
     Cardano.Wallet.Launch.Cluster.ClusterEra
     Cardano.Wallet.Launch.Cluster.ClusterM
+    Cardano.Wallet.Launch.Cluster.CommandLine
     Cardano.Wallet.Launch.Cluster.Config
     Cardano.Wallet.Launch.Cluster.ConfiguredPool
     Cardano.Wallet.Launch.Cluster.Env
@@ -73,7 +74,6 @@ library
     Cardano.Wallet.Launch.Cluster.StakeCertificates
     Cardano.Wallet.Launch.Cluster.Tx
     Cardano.Wallet.Launch.Cluster.UnsafeInterval
-    Cardano.Wallet.LocalCluster
 
   if flag(release)
     ghc-options: -O2 -Werror
@@ -135,12 +135,24 @@ library
 
 executable local-cluster
   import:         language, opts-exe
-  main-is:        Cluster.hs
+  main-is:        ./local-cluster.hs
   hs-source-dirs: exe
 
   build-depends:
     , base
+    , cardano-addresses
+    , cardano-wallet-application-extras
+    , cardano-wallet-launcher
+    , cardano-wallet-primitive
+    , iohk-monitoring-extra
+    , lens
     , local-cluster
+    , optparse-applicative
+    , pathtype
+    , resourcet
+    , temporary-extra
+    , unliftio
+    , with-utf8
 
 test-suite local-cluster-test
   import:         language, opts-exe
